+++ packages/data-context/src/data/coreDataShape.ts
@@ -1,5 +1,5 @@
 import { FoundBrowser, Editor, AllowedState, AllModeOptions, TestingType, BrowserStatus, PACKAGE_MANAGERS, AuthStateName, MIGRATION_STEPS, MigrationStep, BannerState } from '@packages/types'
-import { WizardBundler, ResolvedComponentFrameworkDefinition, CT_FRAMEWORKS, resolveComponentFrameworkDefinition, solidJs } from '@packages/scaffold-config'
+import { WizardBundler, ResolvedComponentFrameworkDefinition, CT_FRAMEWORKS, resolveComponentFrameworkDefinition } from '@packages/scaffold-config'
 import type { NexusGenObjects } from '@packages/graphql/src/gen/nxs.gen'
 import type { App, BrowserWindow } from 'electron'
 import type { ChildProcess } from 'child_process'
@@ -198,7 +198,8 @@ export function makeCoreData (modeOptions: Partial<AllModeOptions> = {}): CoreDa
       chosenManualInstall: false,
       detectedBundler: null,
       detectedFramework: null,
-      frameworks: [...CT_FRAMEWORKS, solidJs].map((framework) => resolveComponentFrameworkDefinition(framework)),
+      // TODO: API to add third party frameworks to this list.
+      frameworks: CT_FRAMEWORKS.map((framework) => resolveComponentFrameworkDefinition(framework)),
     },
     migration: {
       step: 'renameAuto',
