+++ packages/scaffold-config/test/unit/detect.spec.ts
@@ -3,9 +3,10 @@ import { expect } from 'chai'
 import fs from 'fs-extra'
 import globby from 'globby'
 import type { ProjectFixtureDir } from '@tooling/system-tests'
-import { detectFramework, detectLanguage, PkgJson, CT_FRAMEWORKS } from '../../src'
+import { detectFramework, detectLanguage, PkgJson, CT_FRAMEWORKS, resolveComponentFrameworkDefinition, WIZARD_DEPENDENCY_WEBPACK } from '../../src'
 import Fixtures from '@tooling/system-tests'
 import path from 'path'
+import { solidJs, solidDep } from './fixtures'
 
 beforeEach(() => {
   // @ts-ignore
@@ -71,12 +72,14 @@ function fakeDepsInNodeModules (cwd: string, deps: Array<DepToFake | DevDepToFak
   }
 }
 
+const resolvedCtFrameworks = CT_FRAMEWORKS.map((x) => resolveComponentFrameworkDefinition(x))
+
 describe('detectFramework', () => {
   it('Create React App v4', async () => {
     const projectPath = await scaffoldMigrationProject('create-react-app-unconfigured')
 
     fakeDepsInNodeModules(projectPath, [{ dependency: 'react-scripts', version: '5.0.0' }])
-    const actual = await detectFramework(projectPath, CT_FRAMEWORKS)
+    const actual = await detectFramework(projectPath, resolvedCtFrameworks)
 
     expect(actual.framework?.type).to.eq('reactscripts')
   })
@@ -85,7 +88,7 @@ describe('detectFramework', () => {
     const projectPath = await scaffoldMigrationProject('create-react-app-unconfigured')
 
     fakeDepsInNodeModules(projectPath, [{ dependency: 'react-scripts', version: '4.0.0' }])
-    const actual = await detectFramework(projectPath, CT_FRAMEWORKS)
+    const actual = await detectFramework(projectPath, resolvedCtFrameworks)
 
     expect(actual.framework?.type).to.eq('reactscripts')
   })
@@ -98,7 +101,7 @@ describe('detectFramework', () => {
       { devDependency: 'webpack', version: '5.0.0' },
     ])
 
-    const actual = await detectFramework(projectPath, CT_FRAMEWORKS)
+    const actual = await detectFramework(projectPath, resolvedCtFrameworks)
 
     expect(actual.framework?.type).to.eq('react')
     expect(actual.bundler?.type).to.eq('webpack')
@@ -112,7 +115,7 @@ describe('detectFramework', () => {
       { dependency: 'vue', version: '2.5.0' },
     ])
 
-    const actual = await detectFramework(projectPath, CT_FRAMEWORKS)
+    const actual = await detectFramework(projectPath, resolvedCtFrameworks)
 
     expect(actual.framework?.type).to.eq('vueclivue2')
     expect(actual.bundler?.type).to.eq('webpack')
@@ -126,7 +129,7 @@ describe('detectFramework', () => {
       { dependency: 'vue', version: '3.2.0' },
     ])
 
-    const actual = await detectFramework(projectPath, CT_FRAMEWORKS)
+    const actual = await detectFramework(projectPath, resolvedCtFrameworks)
 
     expect(actual.framework?.type).to.eq('vueclivue3')
     expect(actual.bundler?.type).to.eq('webpack')
@@ -140,7 +143,7 @@ describe('detectFramework', () => {
       { dependency: 'vue', version: '3.2.0' },
     ])
 
-    const actual = await detectFramework(projectPath, CT_FRAMEWORKS)
+    const actual = await detectFramework(projectPath, resolvedCtFrameworks)
 
     expect(actual.framework?.type).to.eq('vueclivue3')
     expect(actual.bundler?.type).to.eq('webpack')
@@ -154,7 +157,7 @@ describe('detectFramework', () => {
       { dependency: 'react', version: '17.0.0' },
     ])
 
-    const actual = await detectFramework(projectPath, CT_FRAMEWORKS)
+    const actual = await detectFramework(projectPath, resolvedCtFrameworks)
 
     expect(actual.framework?.type).to.eq('react')
     expect(actual.bundler?.type).to.eq('vite')
@@ -168,7 +171,7 @@ describe('detectFramework', () => {
       { dependency: 'react', version: '17.0.0' },
     ])
 
-    const actual = await detectFramework(projectPath, CT_FRAMEWORKS)
+    const actual = await detectFramework(projectPath, resolvedCtFrameworks)
 
     expect(actual.framework?.type).to.eq('react')
     expect(actual.bundler?.type).to.eq('vite')
@@ -182,7 +185,7 @@ describe('detectFramework', () => {
       { dependency: 'vue', version: '3.0.0' },
     ])
 
-    const actual = await detectFramework(projectPath, CT_FRAMEWORKS)
+    const actual = await detectFramework(projectPath, resolvedCtFrameworks)
 
     expect(actual.framework?.type).to.eq('vue3')
     expect(actual.bundler?.type).to.eq('vite')
@@ -197,7 +200,7 @@ describe('detectFramework', () => {
         { dependency: 'next', version: v },
       ])
 
-      const actual = await detectFramework(projectPath, CT_FRAMEWORKS)
+      const actual = await detectFramework(projectPath, resolvedCtFrameworks)
 
       expect(actual.framework?.type).to.eq('nextjs')
       expect(actual.bundler?.type).to.eq('webpack')
@@ -212,7 +215,7 @@ describe('detectFramework', () => {
         { dependency: '@angular/cli', version: v },
       ])
 
-      const actual = await detectFramework(projectPath, CT_FRAMEWORKS)
+      const actual = await detectFramework(projectPath, resolvedCtFrameworks)
 
       expect(actual.framework?.type).to.eq('angular')
       expect(actual.bundler?.type).to.eq('webpack')
@@ -228,7 +231,7 @@ describe('detectFramework', () => {
         { dependency: 'vite', version: v },
       ])
 
-      const actual = await detectFramework(projectPath, CT_FRAMEWORKS)
+      const actual = await detectFramework(projectPath, resolvedCtFrameworks)
 
       expect(actual.framework?.type).to.eq('svelte')
       expect(actual.bundler?.type).to.eq('vite')
@@ -243,7 +246,7 @@ describe('detectFramework', () => {
       { dependency: 'webpack', version: '5.0.0' },
     ])
 
-    const actual = await detectFramework(projectPath, CT_FRAMEWORKS)
+    const actual = await detectFramework(projectPath, resolvedCtFrameworks)
 
     expect(actual.framework?.type).to.eq('svelte')
     expect(actual.bundler?.type).to.eq('webpack')
@@ -257,7 +260,7 @@ describe('detectFramework', () => {
     // monorepo like situations where there can be multiple levels of
     // node_modules above the projectPath.
     fs.rmSync(path.join(Fixtures.cyTmpDir, 'node_modules'), { recursive: true, force: true })
-    const actual = await detectFramework(projectPath, CT_FRAMEWORKS)
+    const actual = await detectFramework(projectPath, resolvedCtFrameworks)
 
     expect(actual.framework).to.be.undefined
     expect(actual.bundler).to.be.undefined
@@ -466,5 +469,33 @@ describe('detectLanguage', () => {
 
       expect(actual).to.eq('ts')
     })
+  })
+})
+
+describe('resolveComponentFrameworkDefinition', () => {
+  it('resolves a first party framework', async () => {
+    const projectRoot = await scaffoldMigrationProject('migration-ts-files-only')
+
+    fakeDepsInNodeModules(projectRoot, [
+      { dependency: 'solid-js', version: '1.0.0' },
+      { dependency: 'webpack', version: '4.0.0' },
+    ])
+
+    const result = resolveComponentFrameworkDefinition(solidJs)
+
+    expect(await result.dependencies('webpack', projectRoot)).to.deep.include.members([
+      {
+        dependency: solidDep,
+        detectedVersion: '1.0.0',
+        loc: `${projectRoot}/node_modules/solid-js/package.json`,
+        satisfied: true,
+      },
+      {
+        dependency: WIZARD_DEPENDENCY_WEBPACK,
+        detectedVersion: '4.0.0',
+        loc: `${projectRoot}/node_modules/webpack/package.json`,
+        satisfied: true,
+      },
+    ])
   })
 })
